管理者機能更新

+++ app/controllers/public/favorites_controller.rb
@@ -1,15 +1,15 @@
 class Public::FavoritesController < ApplicationController
+  before_action :authenticate_end_user!
   def create
     post_machineko = PostMachineko.find(params[:post_machineko_id])
-    favorite = current_end_user.favorites.new(post_machineko_id: post_machineko.id)
-    favorite.save
+    current_end_user.favorites.find_or_create_by(post_machineko_id: post_machineko.id)
     redirect_to request.referer #同じページへ遷移
   end
   
   def destroy
     post_machineko = PostMachineko.find(params[:post_machineko_id])
-    favorite =current_end_user.favorites.find_by(post_machineko_id: post_machineko.id)
-    favorite.destroy
+    favorite = current_end_user.favorites.find_by(post_machineko_id: post_machineko.id)
+    favorite.destroy if favorite
     redirect_to request.referer #同じページへ遷移
   end
 end
